servers need a different hostname and tmux prefix
When I'm logged into a server I like to be able to see that I'm on
the server easily. I print the hostname on each prompt to make it
clear.

I also need a different prefix for tmux and chose Ctrl-s for now.

In the future I can try to set up a way to do this without creating
a new branch.